requirements_build.txt
# متطلبات إضافية للبناء والتصدير
pyinstaller>=6.0.0
pillow>=10.0.0
auto-py-to-exe>=2.4.0

# المتطلبات الأساسية للتطبيق
PyQt5==5.15.10
PyQt5-tools==5.15.9.3.3
PyQtWebEngine==5.15.6
python-dotenv==1.0.0
bcrypt==4.0.1
Pillow==10.0.0
reportlab==4.0.4
jinja2==3.1.2
supabase==2.3.4
storage3==0.7.7
arabic-reshaper==3.0.0
python-bidi==0.4.2

# متطلبات إضافية لضمان التوافق
setuptools>=65.0.0
wheel>=0.38.0
pefile>=2023.2.7
